Migrate projects page index, project components, and projects service into new app directory - Split project index into:     - a server page file for routing/data retrieval     - pageClient for client side data handling
